Adding support for variable data length
The length of the data is encoded in the first 3 bytes of the transmission:
  - Byte 0: data length
  - Byte 1: ECC
  - Byte 2: ECC

For this type of transmission it is necessary to add an "end-of-message"
marker at the end of the Tx. This marker is used by the receiver to
understand when the transmission is over. At this point the receiver starts
analyzing the recorded data. First, the data length is decoded. After, this
length is used to try to decode the remaining data.

The number ECC bytes used for the payload is calculated as 2*(N/5), where N
is the length of the payload.